refactor: env utils
