+++ scripts/src/commands/index.ts
@@ -3,18 +3,20 @@ import { Command } from "commander";
 import { logProgramHelp } from "../utils";
 
 // Commands
+import devCmd from "./dev";
+import docsCmd from "./docs";
 import strapiCmd from "./strapi";
 import wpCmd from "./wp";
-import docsCmd from "./docs";
 
 const program = new Command();
 
 program.version("1.0.0").description("SAMI Website Scripts");
 
 /** add sub-commands from child folders */
+program.addCommand(devCmd);
+program.addCommand(docsCmd);
 program.addCommand(strapiCmd);
 program.addCommand(wpCmd);
-program.addCommand(docsCmd);
 
 if (!process.argv.slice(2).length) {
   logProgramHelp(program);
